% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.me.R
\name{calculate.me}
\alias{calculate.me}
\title{calculate.me function}
\usage{
calculate.me(
  bfileprefix,
  method,
  plink_path = NULL,
  gear_path = NULL,
  freq_path = NULL,
  pop_size = NULL
)
}
\arguments{
\item{bfileprefix}{bfileprefix of your plink bfiles.}

\item{method}{method = "LB": a approximation approach. LB method is applied using "GEAR". The default interation time is 100.
method = "GRM": based on 1/var(grm.off.diag); GRM method is applied using "plink"
We highly suggest using method = "LB" for biobank-scale data.}

\item{plink_path}{plink path}

\item{gear_path}{gear path}

\item{freq_path}{frequency file (.freq) path}

\item{pop_size}{population size}
}
\value{
me
}
\description{
calculate.me function
}
\examples{
\dontrun{calculate.me(bfileprefix = "1KG-EUR.qc.inter", method = "LB")}
}
