% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepKin.classification.R
\name{deepKin.classification}
\alias{deepKin.classification}
\title{deepKin.classification}
\usage{
deepKin.classification(theta.x, me, alpha)
}
\arguments{
\item{theta.x}{input a vector of theta.x to be tested}

\item{me}{the number of effective markers}

\item{alpha}{significant level}
}
\value{
a data frame with theta, logpt, logpt1, class, quality
}
\description{
deepKin.classification
}
\examples{
\dontrun{qc = kingx[which(kingx$king>theta.min), "king"]
                   qc.rst = deepKin.classification(theta.x = qc, me = me, alpha = alpha)}
}
