% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepKin.R
\name{deepKin}
\alias{deepKin}
\title{deepKin function
This returns a list of deepKin results based on three principles, including the suggested threshold for each degree of relatedness}
\usage{
deepKin(n, me, alpha = 0.05, beta = 0.1, max.degree = 5)
}
\arguments{
\item{n}{The population size}

\item{me}{The effective number of markers}

\item{alpha}{Type I error rate}

\item{beta}{Type II error rate}

\item{max.degree}{User specified deepest interesting degree}
}
\value{
a list
}
\description{
deepKin function
This returns a list of deepKin results based on three principles, including the suggested threshold for each degree of relatedness
}
\examples{
deepKin(n = 1000, me = 300, alpha = 0.05, beta = 0.1, max.degree = 5)
}
