% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepKin.summary.R
\name{deepKin.summary}
\alias{deepKin.summary}
\title{deepKin.summary function
This returns the summary report of deepKin three principles, including the suggested threshold for each degree of relatedness}
\usage{
deepKin.summary(deepKin)
}
\arguments{
\item{deepKin}{the output of deepKin()}
}
\value{
a string
}
\description{
deepKin.summary function
This returns the summary report of deepKin three principles, including the suggested threshold for each degree of relatedness
}
\examples{
\dontrun{deepKin.summary(deepKin)}
}
