% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{deepKin_power}
\alias{deepKin_power}
\title{power function}
\usage{
deepKin_power(me, theta, alpha)
}
\arguments{
\item{me}{The effective number of markers}

\item{theta}{Target relatedness}

\item{alpha}{Type I error rate, usually sets to 0.05/N after Bonferroni correction}
}
\value{
power
}
\description{
power function
}
\examples{
deepKin_power(me = 5000, theta = 0.5, alpha = 0.05/100000)
}
