% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepTheta.R
\name{deepTheta}
\alias{deepTheta}
\title{deepTheta function
This function is related to the second principle in deepKin analysis.
This function calculates the deepest relatedness supported by the data, based on me, type I and type II error rates}
\usage{
deepTheta(me, alpha, beta)
}
\arguments{
\item{me}{The effective number of markers}

\item{alpha}{Type I error rate}

\item{beta}{Type II error rate}
}
\value{
deepTheta relatedness score
}
\description{
deepTheta function
This function is related to the second principle in deepKin analysis.
This function calculates the deepest relatedness supported by the data, based on me, type I and type II error rates
}
\examples{
deepTheta(me = 5000, alpha = 0.05, beta = 0.1)
}
