% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.indi.id.R
\name{extract.indi.id}
\alias{extract.indi.id}
\title{extract.indi.id function
This function will return individual ID based on index in deepKin results and individual IDs}
\usage{
extract.indi.id(id, index, xcohort, pop_size1 = NULL, pop_size2 = NULL)
}
\arguments{
\item{id}{individual IDs}

\item{index}{index in deepKin results}

\item{xcohort}{xcohort = T for cross-cohort GRM extraction; xcohort = F for one cohort GRM;
Make sure that when you use xcohort = T, the individual order is organized just the same order cohort by cohort.}

\item{pop_size1}{population size for the first cohort when xcohort = T}

\item{pop_size2}{population size for the second cohort when xcohort = T}
}
\value{
two columns, the first one is ID1 and the second column is ID2
}
\description{
extract.indi.id function
This function will return individual ID based on index in deepKin results and individual IDs
}
\examples{
\dontrun{deepkin.qc.id = extract.indi.id(id = id, index = index, xcohort = F)}
}
