% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.plink.grm.R
\name{extract.plink.grm}
\alias{extract.plink.grm}
\title{extract.plink.grm function}
\usage{
extract.plink.grm(bfileprefix, xcohort = F, pop_size1, pop_size2)
}
\arguments{
\item{bfileprefix}{plink bfile prefix}

\item{xcohort}{xcohort = T for cross-cohort GRM extraction; xcohort = F for one cohort GRM;
Make sure that when you use xcohort = T, the individual order is organized just the same order cohort by cohort.}

\item{pop_size1}{population size; population size for the first cohort when xcohort = T}

\item{pop_size2}{population size for the second cohort when xcohort = T}
}
\value{
a list
        $diag the diagnal elements, it is an n vector
        $tri the triangle elements, when xcohort = F, $tri is an n*(n-1) vector; when xcohort = T, it is an n1\*n2 vector
}
\description{
extract.plink.grm function
}
\examples{
\dontrun{grm.rst = extract.plink.grm(bfileprefix, xcohort = F, pop_size1 = n)
                   grm.diag = grm.rst$diag
                   grm.tri  = grm.rst$tri}
}
