% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPlinkGRM.R
\name{extractPlinkGRM}
\alias{extractPlinkGRM}
\title{extractPlinkGRM function}
\usage{
extractPlinkGRM(bfileprefix, xcohort = F, pop_size, pop_size2)
}
\arguments{
\item{bfileprefix}{plink bfile prefix}

\item{xcohort}{xcohort = T for cross-cohort GRM extraction; xcohort = F for one cohort GRM;
Make sure that when you use xcohort = T, the individual order is organized just the same order cohort by cohort.}

\item{pop_size}{population size; population size for the first cohort when xcohort = T}

\item{pop_size2}{population size for the second cohort when xcohort = T}
}
\value{
a list $diag the diagnal elements, $tri the triangle elements
}
\description{
extractPlinkGRM function
}
\examples{
\dontrun{grm.rst = extractPlinkGRM(bfileprefix, xcohort = F, pop_size = n)
                   grm.diag = grm.rst$diag
                   grm.tri  = grm.rst$tri}
}
