% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logalpha.R
\name{logalpha}
\alias{logalpha}
\title{logalpha function
This function is related to the third principle in deepKin analysis
This function calculates the type I error rate based on me, target relatedness, and type II error rate.}
\usage{
logalpha(me, theta, beta)
}
\arguments{
\item{me}{The effective number of markers}

\item{theta}{Target relatedness}

\item{beta}{Type II error rate}
}
\value{
log10 alpha
}
\description{
logalpha function
This function is related to the third principle in deepKin analysis
This function calculates the type I error rate based on me, target relatedness, and type II error rate.
}
\examples{
logalpha(me = 5000, theta = 0.5, beta = 0.1)
}
