% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logp.R
\name{logp}
\alias{logp}
\title{logp function}
\usage{
logp(x, null.theta, me)
}
\arguments{
\item{x}{test theta vector}

\item{null.theta}{the theta of null hypothesis}

\item{me}{the number of effective markers}
}
\value{
logp
}
\description{
logp function
}
\examples{
logp(x = seq(0.3,0.4,0.01), null.theta = 0.25, me = 1000)
}
