% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logpThreshold.R
\name{logpThreshold}
\alias{logpThreshold}
\title{logpThreshold function
This function is related to the third principle in deepKin analysis
This function calculates the p threshold or log10 p threshold based on me, target relatedness, and type II error rate.}
\usage{
logpThreshold(me, theta, beta)
}
\arguments{
\item{me}{The effective number of markers}

\item{theta}{Target relatedness}

\item{beta}{Type II error rate}
}
\value{
log10 p threshold
}
\description{
logpThreshold function
This function is related to the third principle in deepKin analysis
This function calculates the p threshold or log10 p threshold based on me, target relatedness, and type II error rate.
}
\examples{
logpThreshold(me = 5000, theta = 0.5, beta = 0.1)
}
