% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me.min.R
\name{me.min}
\alias{me.min}
\title{me.min function
This function is related to the first principle in deepKin analysis
This function calculates the minimum size of m_e (the effective number of markers) based on
target relatedness, type I and type II error rates.}
\usage{
me.min(theta, alpha, beta)
}
\arguments{
\item{theta}{Target relatedness}

\item{alpha}{Type I error rate}

\item{beta}{Type II error rate}
}
\value{
me.min the minimum size of m_e
}
\description{
me.min function
This function is related to the first principle in deepKin analysis
This function calculates the minimum size of m_e (the effective number of markers) based on
target relatedness, type I and type II error rates.
}
\examples{
me.min(theta = 0.5, alpha = 0.05, beta = 0.1)
}
