% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minMe.R
\name{minMe}
\alias{minMe}
\title{minMe function
This function is related to the first principle in deepKin analysis
This function calculates the minimum size of m_e (the effective number of markers) based on
target relatedness, type I and type II error rates.}
\usage{
minMe(theta, alpha, beta)
}
\arguments{
\item{theta}{Target relatedness}

\item{alpha}{Type I error rate}

\item{beta}{Type II error rate}
}
\value{
minMe the minimum size of m_e
}
\description{
minMe function
This function is related to the first principle in deepKin analysis
This function calculates the minimum size of m_e (the effective number of markers) based on
target relatedness, type I and type II error rates.
}
\examples{
minMe(theta = 0.5, alpha = 0.05, beta = 0.1)
}
