% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.max.R
\name{power.max}
\alias{power.max}
\title{power.max function}
\usage{
power.max(me, theta, alpha)
}
\arguments{
\item{me}{The effective number of markers}

\item{theta}{Target relatedness}

\item{alpha}{Type I error rate, usually sets to 0.05/N after Bonferroni correction}
}
\value{
power.max
}
\description{
power.max function
}
\examples{
power.max(me = 5000, theta = 0.5, alpha = 0.05/100000)
}
