% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta.min.R
\name{theta.min}
\alias{theta.min}
\title{theta.min function
This function calculates the deepest relatedness supported by the data, based on me and siginificant level}
\usage{
theta.min(me, alpha)
}
\arguments{
\item{me}{The effective number of markers}

\item{alpha}{Significant level}
}
\value{
theta.min deepest theta relatedness score
}
\description{
theta.min function
This function calculates the deepest relatedness supported by the data, based on me and siginificant level
}
\examples{
theta.min(me = 5000, alpha = 0.05)
}
