% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepKin.estimation.R
\name{deepKin.estimation}
\alias{deepKin.estimation}
\title{deepKin.estimation function}
\usage{
deepKin.estimation(
  bfileprefix,
  plink_path,
  me,
  xcohort = F,
  pop_size1,
  pop_size2
)
}
\arguments{
\item{bfileprefix}{plink bfile prefix}

\item{plink_path}{path to plink executable file}

\item{me}{the effective number of markers}

\item{xcohort}{xcohort = T for cross-cohort GRM extraction; xcohort = F for one cohort GRM;
Make sure that when you use xcohort = T, the individual order is organized just the same order cohort by cohort.}

\item{pop_size1}{population size for the first cohort when xcohort = T}

\item{pop_size2}{population size for the second cohort when xcohort = T}
}
\value{
a dataframe with two columns
        $theta: deepKin estimates
        $minuslogp: deepKin -log10(p-values)
}
\description{
deepKin.estimation function
}
\examples{
\dontrun{KINGX = deepKin.estimation(grm.diag = grm.diag, grm.tri = grm.tri,
                   xcohort = F, me = me)}
}
