% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepKin.summary.R
\name{deepKin.summary}
\alias{deepKin.summary}
\title{deepKin.summary function
This returns the summary report of deepKin's CRITICAL VALUE of significant relatedness and two guidelines of performing}
\usage{
deepKin.summary(deepKin)
}
\arguments{
\item{deepKin}{the output of deepKin()}
}
\value{
a string
}
\description{
deepKin.summary function
This returns the summary report of deepKin's CRITICAL VALUE of significant relatedness and two guidelines of performing
}
\examples{
\dontrun{deepKin.summary(deepKin)}
}
